OAuth Apps Scoped Grants: refactor to ease future implementation
